[#84] refactor: 포스터 Mapper를 utils 폴더로 이동
